# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a documentation for the Swipe API",
        "title": "Swipe Public API v0.0.1",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/inventories": {
            "get": {
                "description": "get all product from inventory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "parameters": [
                    {
                        "type": "number",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.InvStock"
                        }
                    }
                }
            },
            "put": {
                "description": "update inventory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "parameters": [
                    {
                        "description": "Inventory Request",
                        "name": "inventory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.InvUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            },
            "post": {
                "description": "add product to inventories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "parameters": [
                    {
                        "description": "Inventories Request",
                        "name": "InvDetail",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.InvDetail"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete inventory by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "inventory id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            }
        },
        "/inventories/details": {
            "get": {
                "description": "get product availability in inventories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "parameters": [
                    {
                        "type": "number",
                        "description": "inventory id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.Inventory"
                        }
                    }
                }
            }
        },
        "/inventories/image": {
            "put": {
                "description": "update inventory image",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "parameters": [
                    {
                        "type": "file",
                        "description": "stock image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            }
        },
        "/inventories/specs": {
            "post": {
                "description": "create new specification for inventory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "parameters": [
                    {
                        "description": "Specifications inventory Request",
                        "name": "specs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.Specifications"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.OK"
                            }
                        }
                    }
                }
            }
        },
...
        "/products": {
            "get": {
                "description": "get product information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit number of products",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.Slices-dtos_ProductResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "update product information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "description": "Product Information Request",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateProductInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            },
            "post": {
                "description": "create new product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "description": "Product Request",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateProduct"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete product by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "product id",
                        "name": "pid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    }
                }
            }
        },
        "/products/details": {
            "get": {
                "description": "get product details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "type": "number",
                        "description": "product id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ProductDetail"
                        }
                    }
                }
            }
        },
        "/products/img": {
            "post": {
                "description": "insert new product image",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of product",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "image of product",
                        "name": "img",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.OK"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.Error"
                        }
                    }
                }
            }
        },
        "/products/view": {
            "get": {
                "description": "get product view",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "type of product",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.ProductView"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.CreateProduct": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "dtos.Error": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "dtos.InvDetail": {
            "type": "object",
            "required": [
                "available",
                "currency_code",
                "price",
                "product_id"
            ],
            "properties": {
                "available": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "color_img": {
                    "type": "string"
                },
                "currency_code": {
                    "type": "string"
                },
                "image": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "specs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.InvSpecification"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dtos.InvStock": {
            "type": "object",
            "properties": {
                "header": {
                    "$ref": "#/definitions/dtos.StockHeader"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "stock": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.Inventory"
                    }
                }
            }
        },
        "dtos.InvUpdate": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "available": {
                    "type": "string"
                },
                "currency_code": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dtos.Inventory": {
            "type": "object",
            "required": [
                "available",
                "currency_code",
                "price",
                "product_id"
            ],
            "properties": {
                "available": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "color_img": {
                    "type": "string"
                },
                "currency_code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "specs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.InvSpecification"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dtos.OK": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "dtos.ProductRequest": {
            "type": "object",
            "required": [
                "category_id",
                "description",
                "name",
                "price",
                "supplier_id"
            ],
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "specs": {
                    "$ref": "#/definitions/dtos.ProductSpecs"
                },
                "status": {
                    "type": "string"
                },
                "supplier_id": {
                    "type": "integer"
                }
            }
        },
...
        "dtos.Slices-dtos_ProductResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.ProductResponse"
                    }
                }
            }
        },
...
        "dtos.Specifications": {
            "type": "object",
            "required": [
                "inventory_id",
                "ram",
                "ssd"
            ],
            "properties": {
                "inventory_id": {
                    "type": "integer"
                },
                "ram": {
                    "type": "string"
                },
                "ssd": {
                    "type": "string"
                }
            }
        },
...
        "dtos.UpdateProductInfo": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "specs": {
                    "$ref": "#/definitions/dtos.ProductSpecs"
                },
                "status": {
                    "type": "string"
                },
                "supplier_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}